chore: ToolTip handling service
