+++ src/Uno.UI/DirectUI/CoreImports.cs
@@ -1,0 +1,10 @@
+﻿using Microsoft.UI.Input;
+using Microsoft.UI.Xaml.Controls;
+using Windows.System;
+
+namespace DirectUI;
+
+internal static class CoreImports
+{
+	internal static VirtualKeyModifiers Input_GetKeyboardModifiers() => PlatformHelpers.GetKeyboardModifiers();
+}
