+++ src/Uno.UI/DirectUI/FxCallbacks.cs
@@ -1,6 +1,8 @@
 ﻿using System.Runtime.CompilerServices;
 using Microsoft.UI.Xaml;
+using Microsoft.UI.Xaml.Input;
 using Uno.UI.Xaml.Islands;
+using Windows.System;
 
 namespace DirectUI;
 
@@ -14,4 +16,19 @@ internal static class FxCallbacks
 
 	[MethodImpl(MethodImplOptions.AggressiveInlining)]
 	internal static void XamlRoot_RaiseChanged(XamlRoot xamlRoot) => xamlRoot.RaiseChangedEvent();
+
+	[MethodImpl(MethodImplOptions.AggressiveInlining)]
+	internal static bool KeyboardAccelerator_RaiseKeyboardAcceleratorInvoked(
+		KeyboardAccelerator pNativeAccelerator,
+		DependencyObject pElement) =>
+		KeyboardAccelerator.RaiseKeyboardAcceleratorInvoked(pNativeAccelerator, pElement);
+
+	[MethodImpl(MethodImplOptions.AggressiveInlining)]
+	internal static void UIElement_RaiseProcessKeyboardAccelerators(
+		UIElement pUIElement,
+		VirtualKey key,
+		VirtualKeyModifiers keyModifiers,
+		ref bool pHandled,
+		ref bool pHandledShouldNotImpedeTextInput) =>
+		UIElement.RaiseProcessKeyboardAcceleratorsStatic(pUIElement, key, keyModifiers, ref pHandled, ref pHandledShouldNotImpedeTextInput);
 }
