+++ src/Uno.UI/UI/Xaml/Controls/ToolTip/ToolTipService.cs
@@ -14,355 +14,287 @@ using UIKit;
 using AppKit;
 #endif
 
-namespace Controls
+namespace Microsoft.UI.Xaml.Controls;
+
+public partial class ToolTipService
 {
-	partial class ToolTipService
+	private static ToolTip m_CurrentToolTip;
+	private static uint m_LastEnteredFrameId;
+	private static DispatcherTimer m_OpenTimer;
+	private static DispatcherTimer m_CloseTimer;
+
+	private static void OnToolTipChanged(DependencyObject dependencyObject, DependencyPropertyChangedEventArgs e)
 	{
-		#region DependencyProperty: ToolTip
-
-		public static DependencyProperty ToolTipProperty { get; } =
-			DependencyProperty.RegisterAttached(
-				"ToolTip",
-				typeof(object),
-				typeof(ToolTipService),
-				new FrameworkPropertyMetadata(default, OnToolTipChanged));
-
-		public static object GetToolTip(DependencyObject element) => element.GetValue(ToolTipProperty);
-		public static void SetToolTip(DependencyObject element, object value) => element.SetValue(ToolTipProperty, value);
-
-		#endregion
-		#region DependencyProperty: Placement
-
-		public static DependencyProperty PlacementProperty { get; } = DependencyProperty.RegisterAttached(
-			"Placement",
-			typeof(PlacementMode),
-			typeof(ToolTipService),
-			new FrameworkPropertyMetadata(PlacementMode.Top, OnPlacementChanged));
-
-		public static PlacementMode GetPlacement(DependencyObject element) => (PlacementMode)element.GetValue(PlacementProperty);
-		public static void SetPlacement(DependencyObject element, PlacementMode value) => element.SetValue(PlacementProperty, value);
-
-		#endregion
-
-		#region DependencyProperty: PlacementTarget
-
-		/// <summary>
-		/// Gets or sets the object relative to which the tooltip is positioned.
-		/// </summary>
-		public static DependencyProperty PlacementTargetProperty { get; } =
-			DependencyProperty.RegisterAttached(
-				"PlacementTarget",
-				typeof(UIElement),
-				typeof(ToolTipService),
-				new FrameworkPropertyMetadata(default(UIElement)));
-
-		/// <summary>
-		/// Gets the ToolTipService.PlacementTarget XAML attached property value for the specified target element.
-		/// </summary>
-		/// <param name="element">The target element for the attached property value.</param>
-		/// <returns>The visual element that the tooltip is positioned relative to.</returns>
-		public static UIElement GetPlacementTarget(DependencyObject element) => (UIElement)element.GetValue(PlacementTargetProperty);
-
-		/// <summary>
-		/// Sets the ToolTipService.PlacementTarget XAML attached property value for the specified target element.
-		/// </summary>
-		/// <param name="element">The target element for the attached property value.</param>
-		/// <param name="value">The visual element that should be the placement target for the tooltip.</param>
-		public static void SetPlacementTarget(DependencyObject element, UIElement value) => element.SetValue(PlacementTargetProperty, value);
-
-		#endregion
-
-		#region DependencyProperty: ToolTipReference
-
-		internal static DependencyProperty ToolTipReferenceProperty { get; } = DependencyProperty.RegisterAttached(
-			"ToolTipReference",
-			typeof(ToolTip),
-			typeof(ToolTipService),
-			new FrameworkPropertyMetadata(default(ToolTip)));
-
-		internal static ToolTip GetToolTipReference(DependencyObject obj) => (ToolTip)obj.GetValue(ToolTipReferenceProperty);
-		internal static void SetToolTipReference(DependencyObject obj, ToolTip value) => obj.SetValue(ToolTipReferenceProperty, value);
-
-		#endregion
-
-		private static ToolTip m_CurrentToolTip;
-		private static uint m_LastEnteredFrameId;
-		private static DispatcherTimer m_OpenTimer;
-		private static DispatcherTimer m_CloseTimer;
-
-		private static void OnToolTipChanged(DependencyObject dependencyobject, DependencyPropertyChangedEventArgs e)
+		if (!FeatureConfiguration.ToolTip.UseToolTips)
 		{
-			if (!FeatureConfiguration.ToolTip.UseToolTips)
+			return; // ToolTips are disabled
+		}
+
+		if (!(dependencyObject is FrameworkElement owner))
+		{
+			return;
+		}
+
+		if (e.NewValue is null)
+		{
+			DisposePreviousToolTip();
+		}
+		else if (e.NewValue is ToolTip newToolTip)
+		{
+			var previousToolTip = GetToolTipReference(owner);
+
+			// dispose the previous tooltip
+			if (previousToolTip != null && newToolTip != previousToolTip)
 			{
-				return; // ToolTips are disabled
+				DisposePreviousToolTip(previousToolTip);
 			}
 
-			if (!(dependencyobject is FrameworkElement owner))
+			// setup new tooltip
+			if (newToolTip != previousToolTip)
 			{
-				return;
+				SetupToolTip(newToolTip);
 			}
-
-			if (e.NewValue is null)
+		}
+		else
+		{
+			var previousToolTip = GetToolTipReference(owner);
+			if (e.OldValue is ToolTip oldPrevious && oldPrevious == previousToolTip)
 			{
-				DisposePreviousToolTip();
+				// dispose and setup a new tooltip
+				// to avoid corrupting previous tooltip's content with new value
+				DisposePreviousToolTip(previousToolTip);
+				SetupToolTip(new ToolTip { Content = e.NewValue });
 			}
-			else if (e.NewValue is ToolTip newToolTip)
+			else if (previousToolTip != null)
 			{
-				var previousToolTip = GetToolTipReference(owner);
-
-				// dispose the previous tooltip
-				if (previousToolTip != null && newToolTip != previousToolTip)
-				{
-					DisposePreviousToolTip(previousToolTip);
-				}
-
-				// setup new tooltip
-				if (newToolTip != previousToolTip)
-				{
-					SetupToolTip(newToolTip);
-				}
+				// update the old tooltip with new content
+				previousToolTip.Content = e.NewValue;
 			}
 			else
 			{
-				var previousToolTip = GetToolTipReference(owner);
-				if (e.OldValue is ToolTip oldPrevious && oldPrevious == previousToolTip)
-				{
-					// dispose and setup a new tooltip
-					// to avoid corrupting previous tooltip's content with new value
-					DisposePreviousToolTip(previousToolTip);
-					SetupToolTip(new ToolTip { Content = e.NewValue });
-				}
-				else if (previousToolTip != null)
-				{
-					// update the old tooltip with new content
-					previousToolTip.Content = e.NewValue;
-				}
-				else
-				{
-					// setup a new tooltip
-					SetupToolTip(new ToolTip { Content = e.NewValue });
-				}
-			}
-
-			void SetupToolTip(ToolTip toolTip)
-			{
-				toolTip.Placement = GetPlacement(toolTip);
-				toolTip.SetAnchor(GetPlacementTarget(owner) ?? owner);
-
-				SetToolTipReference(owner, toolTip);
-				toolTip.OwnerEventSubscriptions = SubscribeToEvents(owner, toolTip);
-			}
-			void DisposePreviousToolTip(ToolTip toolTip = null)
-			{
-				toolTip ??= GetToolTipReference(owner);
-
-				toolTip.OwnerEventSubscriptions?.Dispose();
-				toolTip.OwnerEventSubscriptions = null;
-				SetToolTipReference(owner, null);
+				// setup a new tooltip
+				SetupToolTip(new ToolTip { Content = e.NewValue });
 			}
 		}
 
-		private static void OnPlacementChanged(DependencyObject dependencyobject, DependencyPropertyChangedEventArgs e)
+		void SetupToolTip(ToolTip toolTip)
 		{
-			if (GetToolTipReference(dependencyobject) is { } tooltip)
-			{
-				tooltip.Placement = (PlacementMode)e.NewValue;
-			}
+			toolTip.Placement = GetPlacement(toolTip);
+			toolTip.SetAnchor(GetPlacementTarget(owner) ?? owner);
+
+			SetToolTipReference(owner, toolTip);
+			toolTip.OwnerEventSubscriptions = SubscribeToEvents(owner, toolTip);
+		}
+		void DisposePreviousToolTip(ToolTip toolTip = null)
+		{
+			toolTip ??= GetToolTipReference(owner);
+
+			toolTip.OwnerEventSubscriptions?.Dispose();
+			toolTip.OwnerEventSubscriptions = null;
+			SetToolTipReference(owner, null);
+		}
+	}
+
+	private static void OnPlacementChanged(DependencyObject dependencyobject, DependencyPropertyChangedEventArgs e)
+	{
+		if (GetToolTipReference(dependencyobject) is { } tooltip)
+		{
+			tooltip.Placement = (PlacementMode)e.NewValue;
+		}
+	}
+
+	private static void OpenToolTipImpl(ToolTip toolTip)
+	{
+		if (m_CurrentToolTip is { })
+		{
+			// Only one instance of the tooltip can be opened at any time.
+			CloseToolTipImpl(m_CurrentToolTip);
 		}
 
-		private static void OpenToolTipImpl(ToolTip toolTip)
+		if (toolTip is { })
 		{
-			if (m_CurrentToolTip is { })
-			{
-				// Only one instance of the tooltip can be opened at any time.
-				CloseToolTipImpl(m_CurrentToolTip);
-			}
+			m_CurrentToolTip = toolTip;
 
-			if (toolTip is { })
-			{
-				m_CurrentToolTip = toolTip;
+			m_OpenTimer.Start();
+			m_CloseTimer?.Stop();
+		}
+	}
 
-				m_OpenTimer.Start();
-				m_CloseTimer?.Stop();
-			}
+	private static void CloseToolTipImpl(ToolTip toolTip)
+	{
+		if (m_CurrentToolTip == toolTip)
+		{
+			m_OpenTimer?.Stop();
+			m_CloseTimer?.Stop();
+
+			m_CurrentToolTip.IsOpen = false;
+		}
+		else
+		{
+			toolTip.IsOpen = false;
+		}
+	}
+
+	private static void OnOpenTimerTick(object sender, object e)
+	{
+		m_OpenTimer.Stop();
+
+		if (m_CurrentToolTip is { })
+		{
+			m_CurrentToolTip.IsOpen = true;
 		}
 
-		private static void CloseToolTipImpl(ToolTip toolTip)
+		if (m_CloseTimer is null)
 		{
-			if (m_CurrentToolTip == toolTip)
-			{
-				m_OpenTimer?.Stop();
-				m_CloseTimer?.Stop();
-
-				m_CurrentToolTip.IsOpen = false;
-			}
-			else
-			{
-				toolTip.IsOpen = false;
-			}
+			m_CloseTimer = new DispatcherTimer();
+			m_CloseTimer.Interval = TimeSpan.FromMilliseconds(FeatureConfiguration.ToolTip.ShowDuration);
+			m_CloseTimer.Tick += OnCloseTimerTick;
 		}
+		m_CloseTimer.Start();
+	}
 
-		private static void OnOpenTimerTick(object sender, object e)
+	private static void OnCloseTimerTick(object sender, object e)
+	{
+		m_CloseTimer.Stop();
+
+		if (m_CurrentToolTip is { })
 		{
-			m_OpenTimer.Stop();
-
-			if (m_CurrentToolTip is { })
-			{
-				m_CurrentToolTip.IsOpen = true;
-			}
-
-			if (m_CloseTimer is null)
-			{
-				m_CloseTimer = new DispatcherTimer();
-				m_CloseTimer.Interval = TimeSpan.FromMilliseconds(FeatureConfiguration.ToolTip.ShowDuration);
-				m_CloseTimer.Tick += OnCloseTimerTick;
-			}
-			m_CloseTimer.Start();
+			m_CurrentToolTip.IsOpen = false;
 		}
+	}
 
-		private static void OnCloseTimerTick(object sender, object e)
+	private static IDisposable SubscribeToEvents(FrameworkElement control, ToolTip toolTip)
+	{
+		// event subscriptions
+		if (control.IsLoaded)
 		{
-			m_CloseTimer.Stop();
-
-			if (m_CurrentToolTip is { })
-			{
-				m_CurrentToolTip.IsOpen = false;
-			}
+			OnOwnerLoaded(control, null);
 		}
+		control.Loaded += OnOwnerLoaded;
+		control.Unloaded += OnOwnerUnloaded;
 
-		private static IDisposable SubscribeToEvents(FrameworkElement control, ToolTip toolTip)
+		return Disposable.Create(() =>
 		{
-			// event subscriptions
-			if (control.IsLoaded)
-			{
-				OnOwnerLoaded(control, null);
-			}
-			control.Loaded += OnOwnerLoaded;
-			control.Unloaded += OnOwnerUnloaded;
+			control.Loaded -= OnOwnerLoaded;
+			control.Unloaded -= OnOwnerUnloaded;
+			OnOwnerUnloaded(control, null);
 
-			return Disposable.Create(() =>
-			{
-				control.Loaded -= OnOwnerLoaded;
-				control.Unloaded -= OnOwnerUnloaded;
-				OnOwnerUnloaded(control, null);
+			CloseToolTipImpl(toolTip);
+		});
+	}
 
+	private static void OnOwnerVisibilityChanged(DependencyObject sender, DependencyProperty dp)
+	{
+		if (sender is FrameworkElement owner && owner.Visibility != Visibility.Visible)
+		{
+			if (GetToolTipReference(owner) is { } toolTip)
+			{
 				CloseToolTipImpl(toolTip);
+			}
+		}
+	}
+
+	private static void OnOwnerLoaded(object sender, RoutedEventArgs e)
+	{
+		if (sender is FrameworkElement owner && GetToolTipReference(owner) is { } toolTip)
+		{
+			owner.PointerEntered += OnPointerEntered;
+			owner.PointerExited += OnPointerExited;
+			owner.Tapped += OnTapped;
+			owner.AddHandler(UIElement.KeyDownEvent, new KeyEventHandler(OnKeyDown), true);
+			if (owner is ButtonBase)
+			{
+				owner.AddHandler(UIElement.PointerPressedEvent, new PointerEventHandler(OnPointerPressed), true);
+			}
+			var token = owner.RegisterPropertyChangedCallback(UIElement.VisibilityProperty, OnOwnerVisibilityChanged);
+			toolTip.OwnerVisibilitySubscription = Disposable.Create(() =>
+			{
+				owner.UnregisterPropertyChangedCallback(UIElement.VisibilityProperty, token);
 			});
 		}
+	}
 
-		private static void OnOwnerVisibilityChanged(DependencyObject sender, DependencyProperty dp)
+	private static void OnOwnerUnloaded(object sender, RoutedEventArgs e)
+	{
+		if (sender is FrameworkElement owner && GetToolTipReference(owner) is { } toolTip)
 		{
-			if (sender is FrameworkElement owner && owner.Visibility != Visibility.Visible)
+			CloseToolTipImpl(toolTip);
+
+			owner.PointerEntered -= OnPointerEntered;
+			owner.PointerExited -= OnPointerExited;
+			owner.Tapped -= OnTapped;
+			owner.AddHandler(UIElement.KeyDownEvent, new KeyEventHandler(OnKeyDown), true);
+			if (owner is ButtonBase)
 			{
-				if (GetToolTipReference(owner) is { } toolTip)
-				{
-					CloseToolTipImpl(toolTip);
-				}
+				owner.RemoveHandler(UIElement.PointerPressedEvent, new PointerEventHandler(OnPointerPressed));
 			}
+			toolTip.OwnerVisibilitySubscription?.Dispose();
+			toolTip.OwnerVisibilitySubscription = null;
 		}
+	}
 
-		private static void OnOwnerLoaded(object sender, RoutedEventArgs e)
+	private static void OnPointerEntered(object sender, PointerRoutedEventArgs e)
+	{
+		// Multiple elements can all receive the same PointerEntered at once (from inner-most to outer-most).
+		// In this case, the inner-most one is the only one that should be shown,
+		// so we are dropping any subsequent events from this frame-id.
+		if (e.FrameId == m_LastEnteredFrameId) return;
+
+		if (sender is FrameworkElement owner && GetToolTipReference(owner) is { } toolTip)
 		{
-			if (sender is FrameworkElement owner && GetToolTipReference(owner) is { } toolTip)
+			if (toolTip.IsOpen) return;
+
+			if (m_OpenTimer is null)
 			{
-				owner.PointerEntered += OnPointerEntered;
-				owner.PointerExited += OnPointerExited;
-				owner.Tapped += OnTapped;
-				owner.AddHandler(UIElement.KeyDownEvent, new KeyEventHandler(OnKeyDown), true);
-				if (owner is ButtonBase)
-				{
-					owner.AddHandler(UIElement.PointerPressedEvent, new PointerEventHandler(OnPointerPressed), true);
-				}
-				var token = owner.RegisterPropertyChangedCallback(UIElement.VisibilityProperty, OnOwnerVisibilityChanged);
-				toolTip.OwnerVisibilitySubscription = Disposable.Create(() =>
-				{
-					owner.UnregisterPropertyChangedCallback(UIElement.VisibilityProperty, token);
-				});
+				m_OpenTimer = new DispatcherTimer();
+				m_OpenTimer.Interval = TimeSpan.FromMilliseconds(FeatureConfiguration.ToolTip.ShowDelay);
+				m_OpenTimer.Tick += OnOpenTimerTick;
 			}
+			m_LastEnteredFrameId = e.FrameId;
+			OpenToolTipImpl(toolTip);
 		}
+	}
 
-		private static void OnOwnerUnloaded(object sender, RoutedEventArgs e)
+	private static void OnPointerExited(object sender, PointerRoutedEventArgs e)
+	{
+		if (sender is FrameworkElement owner && GetToolTipReference(owner) is { } toolTip)
 		{
-			if (sender is FrameworkElement owner && GetToolTipReference(owner) is { } toolTip)
+			CloseToolTipImpl(toolTip);
+		}
+	}
+
+	private static void OnTapped(object sender, TappedRoutedEventArgs e)
+	{
+		if (sender is FrameworkElement owner && GetToolTipReference(owner) is { } toolTip)
+		{
+			CloseToolTipImpl(toolTip);
+		}
+	}
+
+	private static void OnKeyDown(object sender, KeyRoutedEventArgs args)
+	{
+		if (sender is FrameworkElement owner && GetToolTipReference(owner) is { } toolTip)
+		{
+			switch (args.Key)
+			{
+				case VirtualKey.Up:
+				case VirtualKey.Down:
+				case VirtualKey.Left:
+				case VirtualKey.Right:
+					return;
+			}
+
+			CloseToolTipImpl(toolTip);
+		}
+	}
+
+	private static void OnPointerPressed(object sender, PointerRoutedEventArgs e)
+	{
+		if (sender is FrameworkElement owner && GetToolTipReference(owner) is { } toolTip)
+		{
+			if (e.GetCurrentPoint(owner).Properties.IsLeftButtonPressed)
 			{
 				CloseToolTipImpl(toolTip);
-
-				owner.PointerEntered -= OnPointerEntered;
-				owner.PointerExited -= OnPointerExited;
-				owner.Tapped -= OnTapped;
-				owner.AddHandler(UIElement.KeyDownEvent, new KeyEventHandler(OnKeyDown), true);
-				if (owner is ButtonBase)
-				{
-					owner.RemoveHandler(UIElement.PointerPressedEvent, new PointerEventHandler(OnPointerPressed));
-				}
-				toolTip.OwnerVisibilitySubscription?.Dispose();
-				toolTip.OwnerVisibilitySubscription = null;
-			}
-		}
-
-		private static void OnPointerEntered(object sender, PointerRoutedEventArgs e)
-		{
-			// Multiple elements can all receive the same PointerEntered at once (from inner-most to outer-most).
-			// In this case, the inner-most one is the only one that should be shown,
-			// so we are dropping any subsequent events from this frame-id.
-			if (e.FrameId == m_LastEnteredFrameId) return;
-
-			if (sender is FrameworkElement owner && GetToolTipReference(owner) is { } toolTip)
-			{
-				if (toolTip.IsOpen) return;
-
-				if (m_OpenTimer is null)
-				{
-					m_OpenTimer = new DispatcherTimer();
-					m_OpenTimer.Interval = TimeSpan.FromMilliseconds(FeatureConfiguration.ToolTip.ShowDelay);
-					m_OpenTimer.Tick += OnOpenTimerTick;
-				}
-				m_LastEnteredFrameId = e.FrameId;
-				OpenToolTipImpl(toolTip);
-			}
-		}
-
-		private static void OnPointerExited(object sender, PointerRoutedEventArgs e)
-		{
-			if (sender is FrameworkElement owner && GetToolTipReference(owner) is { } toolTip)
-			{
-				CloseToolTipImpl(toolTip);
-			}
-		}
-
-		private static void OnTapped(object sender, TappedRoutedEventArgs e)
-		{
-			if (sender is FrameworkElement owner && GetToolTipReference(owner) is { } toolTip)
-			{
-				CloseToolTipImpl(toolTip);
-			}
-		}
-
-		private static void OnKeyDown(object sender, KeyRoutedEventArgs args)
-		{
-			if (sender is FrameworkElement owner && GetToolTipReference(owner) is { } toolTip)
-			{
-				switch (args.Key)
-				{
-					case VirtualKey.Up:
-					case VirtualKey.Down:
-					case VirtualKey.Left:
-					case VirtualKey.Right:
-						return;
-				}
-
-				CloseToolTipImpl(toolTip);
-			}
-		}
-
-		private static void OnPointerPressed(object sender, PointerRoutedEventArgs e)
-		{
-			if (sender is FrameworkElement owner && GetToolTipReference(owner) is { } toolTip)
-			{
-				if (e.GetCurrentPoint(owner).Properties.IsLeftButtonPressed)
-				{
-					CloseToolTipImpl(toolTip);
-				}
 			}
 		}
 	}
