+++ src/Uno.UI/UI/Xaml/UIElement.cs
@@ -522,7 +522,7 @@ namespace Microsoft.UI.Xaml
 		}
 
 		[NotImplemented]
-		protected internal virtual AutomationPeer OnCreateAutomationPeer() => new AutomationPeer();
+		protected virtual AutomationPeer OnCreateAutomationPeer() => new AutomationPeer();
 
 		internal AutomationPeer OnCreateAutomationPeerInternal() => OnCreateAutomationPeer();
